file descriptors for running_procs
File descriptors and their paths are added in an array of json objects. The array starts
with a special string that should be caught and removed in returner to enable correct
representation on Splunk side.

File descriptors 0, 1, 2 normally correspond to stdin, stdout, and stderr respectively.
Among other things, this can help to track programs receiving input/output from
interactive ttys.