convert relative start time of a process to absolute
By default the start_time attribute of a process is counted in terms of seconds from most
recent reboot. Often the absolute time when the process was started is more of interest
than the relative due to comparisons. This change converts the seconds from the last reboot
into seonds from epoch.