fix login_history query
in certain cases utmp record type 8 is written often to the wtmp file without
tty indicading that it's not correlated with an end of interactive shell.
this commit filters out the utmp type 8 records that do not have tty.
it also increases collecting records from 1h to 1h and 60 seconds to account
for time drift hubble schedule executions can create.